requirements.txt
asgiref==3.2.10 
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-allauth==0.42.0
django-cors-headers==3.5.0
django-filter==2.1.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djongo==1.3.3
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
#psycopg2-binary==2.7.7
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.17.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.2.4
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
